+++ example/lib/pages/basic_prompt.dart
@@ -2,17 +2,18 @@ import 'package:flutter/material.dart';
 import 'package:flutter_scroll_to_top/flutter_scroll_to_top.dart';
 
 class BasicPrompt extends StatelessWidget {
-  final ScrollController scrollController = ScrollController();
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Basic Prompt'),
+        title: const Text('Basic Prompt'),
       ),
       body: ScrollWrapper(
-        scrollController: scrollController,
-        child: ListView.builder(
-          controller: scrollController,
+        builder: (context, properties) => ListView.builder(
+          controller: properties.scrollController,
+          scrollDirection: properties.scrollDirection,
+          reverse: properties.reverse,
+          primary: properties.primary,
           itemBuilder: (context, index) => Padding(
             padding: const EdgeInsets.all(8.0),
             child: ListTile(