+++ example/lib/pages/custom_prompt.dart
@@ -2,23 +2,22 @@ import 'package:flutter/material.dart';
 import 'package:flutter_scroll_to_top/flutter_scroll_to_top.dart';
 
 class CustomPrompt extends StatelessWidget {
-  final ScrollController scrollController = ScrollController();
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Custom Prompt'),
+        title: const Text('Custom Prompt'),
       ),
       body: ScrollWrapper(
-        scrollController: scrollController,
-        promptReplacementBuilder: (BuildContext context, Function function) {
-          return MaterialButton(
-            onPressed: () => function(),
-            child: Text('Scroll to top'),
-          );
-        },
-        child: ListView.builder(
-          controller: scrollController,
+        promptReplacementBuilder: (context, function) => MaterialButton(
+          onPressed: () => function(),
+          child: const Text('Scroll to top'),
+        ),
+        builder: (context, properties) => ListView.builder(
+          controller: properties.scrollController,
+          scrollDirection: properties.scrollDirection,
+          reverse: properties.reverse,
+          primary: properties.primary,
           itemBuilder: (context, index) => Padding(
             padding: const EdgeInsets.all(8.0),
             child: ListTile(