+++ example/lib/pages/nested_scroll_view_example.dart
@@ -12,7 +12,7 @@ class NestedScrollViewExample extends StatelessWidget {
           headerSliverBuilder: (context, innerBoxIsScrolled) => [
             SliverOverlapAbsorber(
               handle: NestedScrollView.sliverOverlapAbsorberHandleFor(context),
-              sliver: SliverSafeArea(
+              sliver: const SliverSafeArea(
                 sliver: SliverAppBar(
                   expandedHeight: 300,
                   collapsedHeight: 100,
@@ -26,17 +26,15 @@ class NestedScrollViewExample extends StatelessWidget {
             builder: (context) {
               NestedScrollView.sliverOverlapAbsorberHandleFor(context);
               return ScrollWrapper(
-                scrollController: scrollController,
-                promptReplacementBuilder:
-                    (BuildContext context, Function function) {
-                  return MaterialButton(
-                    onPressed: () {
-                      function();
-                    },
-                    child: Text('Scroll to top'),
-                  );
-                },
-                child: CustomScrollView(
+                promptReplacementBuilder: (context, function) => MaterialButton(
+                  onPressed: function,
+                  child: const Text('Scroll to top'),
+                ),
+                builder: (context, properties) => CustomScrollView(
+                  controller: properties.scrollController,
+                  scrollDirection: properties.scrollDirection,
+                  reverse: properties.reverse,
+                  primary: properties.primary,
                   slivers: [
                     SliverOverlapInjector(
                       handle: NestedScrollView.sliverOverlapAbsorberHandleFor(