+++ lib/src/scroll_wrapper.dart
@@ -1,21 +1,93 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
+import 'package:flutter/scheduler.dart';
 import 'package:flutter_scroll_to_top/src/ui/expand_animation.dart';
 
 typedef ReplacementBuilder = Widget Function(
     BuildContext context, VoidCallback function);
 
+typedef ScrollBuilder = Widget Function(
+    BuildContext context, ScrollViewProperties properties);
+
 /// Wrap the widget to show a scroll to top prompt over when a certain scroll
 /// offset is reached.
 class ScrollWrapper extends StatefulWidget {
+  const ScrollWrapper({
+    Key? key,
+    required this.builder,
+    this.scrollController,
+    this.scrollDirection = Axis.vertical,
+    bool? primary,
+    this.reverse = false,
+    this.onPromptTap,
+    this.scrollOffsetUntilVisible = 200,
+    this.enabledAtOffset = 200,
+    this.alwaysVisibleAtOffset = false,
+    this.scrollToTopCurve = Curves.fastOutSlowIn,
+    this.scrollToTopDuration = const Duration(milliseconds: 500),
+    this.promptDuration = const Duration(milliseconds: 500),
+    this.promptAnimationCurve = Curves.fastOutSlowIn,
+    this.promptAlignment,
+    this.promptTheme,
+    this.promptAnimationType = PromptAnimation.size,
+    this.promptReplacementBuilder,
+  })  : assert(
+          !(scrollController != null && primary == true),
+          'Primary ScrollViews obtain their ScrollController via inheritance from a PrimaryScrollController widget. '
+          'You cannot both set primary to true and pass an explicit controller.',
+        ),
+        primary = primary ??
+            scrollController == null &&
+                identical(scrollDirection, Axis.vertical),
+        super(key: key);
+
   /// [ScrollController] of the scrollable widget to scroll to the top of when
   /// the prompt is pressed.
-  final ScrollController scrollController;
+  final ScrollController? scrollController;
 
   /// The widget to show the prompt over when scroll is at an offset.
-  final Widget child;
+  final ScrollBuilder builder;
 
-  /// At what scroll offset to show the prompt on.
-  final double promptScrollOffset;
+  /// The axis along which the scroll view scrolls.
+  ///
+  /// Defaults to [Axis.vertical].
+  final Axis scrollDirection;
+
+  /// Whether this is wrapped around a primary scroll view associated with the parent
+  /// [PrimaryScrollController].
+  /// Defaults to true when [scrollDirection] is [Axis.vertical] and
+  /// [controller] is null.
+  final bool primary;
+
+  /// Whether the scroll view scrolls in the reading direction.
+  ///
+  /// For example, if the reading direction is left-to-right and
+  /// [scrollDirection] is [Axis.horizontal], then the scroll view scrolls from
+  /// left to right when [reverse] is false and from right to left when
+  /// [reverse] is true.
+  ///
+  /// Similarly, if [scrollDirection] is [Axis.vertical], then the scroll view
+  /// scrolls from top to bottom when [reverse] is false and from bottom to top
+  /// when [reverse] is true.
+  ///
+  /// Defaults to false.
+  final bool reverse;
+
+  /// Callback function to be executed when the prompt is tapped.
+  final VoidCallback? onPromptTap;
+
+  /// At what scroll offset to enable the prompt on.
+  final double enabledAtOffset;
+
+  /// If the prompt is to be always visible at the provided offset. Setting this
+  /// to false only shows the prompt when the user starts scrolling upwards.
+  /// Default value is false.
+  final bool alwaysVisibleAtOffset;
+
+  /// If [alwaysVisibleAtOffset] is false, what offset should the user scroll in
+  /// the opposite direction (ex, upwards scroll on a non-reversed vertical
+  /// ScrollView) before the prompt becomes visible.
+  final double scrollOffsetUntilVisible;
 
   /// **Replace the prompt button with your own custom widget. Returns the**
   /// **[BuildContext] and the [Function] to call to scroll to top.**
@@ -37,7 +109,7 @@ class ScrollWrapper extends StatefulWidget {
   final PromptButtonTheme? promptTheme;
 
   /// Where on the widget to align the prompt.
-  final Alignment promptAlignment;
+  final Alignment? promptAlignment;
 
   /// Duration it takes for the prompt to come into view/vanish.
   final Duration promptDuration;
@@ -55,47 +127,126 @@ class ScrollWrapper extends StatefulWidget {
   /// Default is [PromptAnimation.size].
   final PromptAnimation promptAnimationType;
 
-  ScrollWrapper({
-    Key? key,
-    required this.scrollController,
-    this.promptScrollOffset = 200,
-    required this.child,
-    this.scrollToTopCurve = Curves.fastOutSlowIn,
-    this.scrollToTopDuration = const Duration(milliseconds: 500),
-    this.promptDuration = const Duration(milliseconds: 500),
-    this.promptAnimationCurve = Curves.fastOutSlowIn,
-    this.promptAlignment = Alignment.topRight,
-    this.promptTheme,
-    this.promptAnimationType = PromptAnimation.size,
-    this.promptReplacementBuilder,
-  }) : super(key: key);
-
   @override
   _ScrollWrapperState createState() => _ScrollWrapperState();
 }
 
 class _ScrollWrapperState extends State<ScrollWrapper> {
-  late PromptButtonTheme promptTheme;
+  late PromptButtonTheme _promptTheme;
+  late ScrollController _scrollController;
+  late Alignment _promptAlignment;
 
   @override
   void initState() {
     super.initState();
-    promptTheme = widget.promptTheme ?? PromptButtonTheme();
+    _promptTheme = widget.promptTheme ?? const PromptButtonTheme();
+    _promptAlignment = widget.promptAlignment ?? _setPromptAlignment();
+  }
+
+  Alignment _setPromptAlignment() {
+    if (identical(widget.scrollDirection, Axis.vertical)) {
+      if (widget.reverse) {
+        return Alignment.bottomRight;
+      } else {
+        return Alignment.topRight;
+      }
+    } else {
+      if (widget.reverse) {
+        return Alignment.topRight;
+      } else {
+        return Alignment.topLeft;
+      }
+    }
+  }
+
+  IconData _getDefaultIcon() {
+    if (identical(widget.scrollDirection, Axis.vertical)) {
+      if (widget.reverse) {
+        return Icons.keyboard_arrow_down_rounded;
+      } else {
+        return Icons.keyboard_arrow_up_rounded;
+      }
+    } else {
+      if (widget.reverse) {
+        return Icons.keyboard_arrow_right_rounded;
+      } else {
+        return Icons.keyboard_arrow_left_rounded;
+      }
+    }
+  }
+
+  @override
+  void didChangeDependencies() {
+    _setupScrollController();
+    _setupListener();
+    super.didChangeDependencies();
   }
 
   @override
   void didUpdateWidget(covariant ScrollWrapper oldWidget) {
     super.didUpdateWidget(oldWidget);
     if (oldWidget.promptTheme != widget.promptTheme) {
-      promptTheme = widget.promptTheme ?? PromptButtonTheme();
+      _promptTheme = widget.promptTheme ?? const PromptButtonTheme();
+    }
+    if (oldWidget.promptAlignment != widget.promptAlignment ||
+        oldWidget.primary != widget.primary ||
+        oldWidget.scrollDirection != widget.scrollDirection) {
+      _promptAlignment = widget.promptAlignment ?? _setPromptAlignment();
     }
   }
 
-  bool scrollTopAtOffset = false;
+  void _setupScrollController() {
+    if (widget.primary) {
+      _scrollController =
+          PrimaryScrollController.of(context) ?? ScrollController();
+    } else {
+      _scrollController = widget.scrollController ?? ScrollController();
+    }
+  }
 
-  void scrollToTop() {
-    widget.scrollController.animateTo(
-      widget.scrollController.position.minScrollExtent,
+  double? _currentScrollStartOffset;
+
+  void _setupListener() {
+    _scrollController.addListener(() {
+      final direction = _scrollController.position.userScrollDirection;
+      if (widget.alwaysVisibleAtOffset) {
+        _checkState();
+      } else if (direction == ScrollDirection.forward) {
+        _currentScrollStartOffset =
+            _currentScrollStartOffset ?? _scrollController.offset;
+        if (_currentScrollStartOffset! - _scrollController.offset >
+            widget.scrollOffsetUntilVisible) {
+          _checkState();
+        }
+      } else {
+        setState(() {
+          _scrollTopAtOffset = false;
+        });
+        _currentScrollStartOffset = null;
+      }
+    });
+  }
+
+  void _checkState() {
+    if (_scrollController.offset > widget.enabledAtOffset &&
+        !_scrollTopAtOffset) {
+      setState(() {
+        _scrollTopAtOffset = true;
+      });
+    } else if (_scrollController.offset <= widget.enabledAtOffset &&
+        _scrollTopAtOffset) {
+      setState(() {
+        _scrollTopAtOffset = false;
+      });
+    }
+  }
+
+  bool _scrollTopAtOffset = false;
+
+  void _scrollToTop() {
+    widget.onPromptTap?.call();
+    _scrollController.animateTo(
+      _scrollController.position.minScrollExtent,
       duration: widget.scrollToTopDuration,
       curve: widget.scrollToTopCurve,
     );
@@ -103,56 +254,44 @@ class _ScrollWrapperState extends State<ScrollWrapper> {
 
   @override
   Widget build(BuildContext context) {
-    Widget child() {
-      return NotificationListener<ScrollNotification>(
-        onNotification: (notification) {
-          if (notification.metrics.pixels > widget.promptScrollOffset &&
-              !scrollTopAtOffset)
-            setState(() {
-              scrollTopAtOffset = true;
-            });
-          else if (notification.metrics.pixels <= widget.promptScrollOffset &&
-              scrollTopAtOffset)
-            setState(() {
-              scrollTopAtOffset = false;
-            });
-          return true;
-        },
-        child: widget.child,
-      );
-    }
-
     return Stack(
       children: [
-        child(),
+        widget.builder(
+          context,
+          ScrollViewProperties(
+              reverse: widget.reverse,
+              primary: widget.primary,
+              scrollDirection: widget.scrollDirection,
+              scrollController: _scrollController),
+        ),
         Align(
-          alignment: widget.promptAlignment,
-          child: SizeExpandedSection(
-            expand: scrollTopAtOffset,
+          alignment: _promptAlignment,
+          child: AnimatePrompt(
+            expand: _scrollTopAtOffset,
             animType: widget.promptAnimationType,
             duration: widget.promptDuration,
             curve: widget.promptAnimationCurve,
-            alignment: widget.promptAlignment,
+            alignment: _promptAlignment,
             child: widget.promptReplacementBuilder != null
-                ? widget.promptReplacementBuilder!(context, scrollToTop)
+                ? widget.promptReplacementBuilder!(context, _scrollToTop)
                 : Padding(
-                    padding: promptTheme.padding,
+                    padding: _promptTheme.padding,
                     child: Material(
-                      elevation: promptTheme.elevation ??
+                      elevation: _promptTheme.elevation ??
                           Theme.of(context).appBarTheme.elevation ??
                           4,
                       clipBehavior: Clip.antiAlias,
                       type: MaterialType.circle,
-                      color: promptTheme.color ??
+                      color: _promptTheme.color ??
                           Theme.of(context).appBarTheme.backgroundColor ??
                           Theme.of(context).primaryColor,
                       child: InkWell(
-                        onTap: scrollToTop,
+                        onTap: _scrollToTop,
                         child: Padding(
-                          padding: promptTheme.iconPadding,
-                          child: promptTheme.icon ??
+                          padding: _promptTheme.iconPadding,
+                          child: _promptTheme.icon ??
                               Icon(
-                                Icons.keyboard_arrow_up_rounded,
+                                _getDefaultIcon(),
                                 color: Theme.of(context)
                                         .appBarTheme
                                         .foregroundColor ??
@@ -170,6 +309,15 @@ class _ScrollWrapperState extends State<ScrollWrapper> {
 }
 
 class PromptButtonTheme {
+  /// Custom prompt button theme to be given to a [ScrollWrapper].
+  const PromptButtonTheme({
+    this.padding = const EdgeInsets.symmetric(horizontal: 16, vertical: 16),
+    this.icon,
+    this.iconPadding = const EdgeInsets.all(8.0),
+    this.elevation,
+    this.color,
+  });
+
   /// Padding around the prompt button.
   final EdgeInsets padding;
 
@@ -185,15 +333,21 @@ class PromptButtonTheme {
   /// Color of the prompt button. Defaults to the accent color of the current
   /// context's [Theme].
   final Color? color;
+}
 
-  /// Custom prompt button theme to be given to a [ScrollWrapper].
-  PromptButtonTheme({
-    this.padding = const EdgeInsets.symmetric(horizontal: 16, vertical: 16),
-    this.icon,
-    this.iconPadding = const EdgeInsets.all(8.0),
-    this.elevation,
-    this.color,
-  });
+class ScrollViewProperties {
+  const ScrollViewProperties(
+      {required this.reverse,
+      required this.primary,
+      required this.scrollDirection,
+      required ScrollController scrollController})
+      : _scrollController = scrollController;
+  final ScrollController? _scrollController;
+  final bool reverse;
+  final bool primary;
+  final Axis scrollDirection;
+
+  ScrollController? get scrollController => primary ? null : _scrollController;
 }
 
 enum PromptAnimation { fade, scale, size }