+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [2.0.0] - 21 November 2021.
+- ***BREAKING CHANGE*** - Replaced `child` parameter with `builder`.
+- ***BREAKING CHANGE*** - Renamed `promptScrollOffset` to `enabledAtOffset`.
+- Added `scrollDirection`, `primary`, `reverse`, `scrollOffsetUntilVisible` and `alwaysVisibleAtOffset` parameters.
+- Added `onPromptTap` callback.
+- Made `scrollController` parameter nullable and added `PrimaryScrollController.of(context)` fallback.
+- Better behaviour for horizontal ScrollViews.
+- Changed default behaviour to show prompt on upwards scroll. Can be disabled by setting `alwaysVisibleAtOffset` to false.
+- Made `ScrollWrapper` and `PromptButtonTheme` constructors const.
+
 ## [1.2.0] - 21 November 2021.
 
 - Added elevation to `PromptButtonTheme`